BugFix: No null comparison for related edge subq
OrientDB has an issue with using the CONTAINS operator with a null
value. Likely because it defaults to the = operator for the element
against each element in the collection.

To avoid this, remove the deletedAt = null comparison from subqueries
that are after related edge queries (these will always produce a
collection)

Also fixed jsdoc issues