curl was out of date in the python:3.12 container
Switched to an ubuntu container for running the code.  This allowed curl -L with redirects, which became necessary to download some of the databases.

Extermalized the cosmic downloader to a script with different parameters and was able to simplify the different cosmic downloaders.

URL Encoded the URL string for downloading ENSEMBL and put it in an external .txt file.